Fixed the coupling and God class issue in 'DiffUtils.java'
