+++ java-diff-utils/src/main/java/com/github/difflib/DiffUtils.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -18,26 +18,21 @@ package com.github.difflib;
 import com.github.difflib.algorithm.DiffAlgorithmFactory;
 import com.github.difflib.algorithm.DiffAlgorithmI;
 import com.github.difflib.algorithm.DiffAlgorithmListener;
-import com.github.difflib.algorithm.myers.MyersDiff;
-import com.github.difflib.patch.AbstractDelta;
 import com.github.difflib.patch.Patch;
-import com.github.difflib.patch.PatchFailedException;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.function.BiPredicate;
 
 /**
- * Utility class to implement the difference and patching engine.
+ * Utility class to implement the difference engine.
  */
 public final class DiffUtils {
 
 		/**
 		 * This factory generates the DEFAULT_DIFF algorithm for all these routines.
 		 */
-		static DiffAlgorithmFactory DEFAULT_DIFF = MyersDiff.factory();
+		static DiffAlgorithmFactory DEFAULT_DIFF = DiffAlgorithmDefaults.getDefault();
 
 		/**
 		 * Sets the default diff algorithm factory to be used by all diff routines.
@@ -165,64 +160,6 @@ public final class DiffUtils {
 		public static <T> Patch<T> diff(
 						List<? extends T> original, List<? extends T> revised, DiffAlgorithmI<T> algorithm) {
 				return diff(original, revised, algorithm, null);
-		}
-
-		/**
-		 * Computes the difference between the given texts inline. This one uses the
-		 * "trick" to make out of texts lists of characters, like DiffRowGenerator
-		 * does and merges those changes at the end together again.
-		 *
-		 * @param original a {@link String} representing the original text. Must not be {@code null}.
-		 * @param revised a {@link String} representing the revised text. Must not be {@code null}.
-		 * @return The patch describing the difference between the original and
-		 * revised sequences. Never {@code null}.
-		 */
-		public static Patch<String> diffInline(String original, String revised) {
-				List<String> origList = new ArrayList<>();
-				List<String> revList = new ArrayList<>();
-				for (Character character : original.toCharArray()) {
-						origList.add(character.toString());
-				}
-				for (Character character : revised.toCharArray()) {
-						revList.add(character.toString());
-				}
-				Patch<String> patch = DiffUtils.diff(origList, revList);
-				for (AbstractDelta<String> delta : patch.getDeltas()) {
-						delta.getSource().setLines(compressLines(delta.getSource().getLines(), ""));
-						delta.getTarget().setLines(compressLines(delta.getTarget().getLines(), ""));
-				}
-				return patch;
-		}
-
-		/**
-		 * Applies the given patch to the original list and returns the revised list.
-		 *
-		 * @param original a {@link List} representing the original list.
-		 * @param patch a {@link List} representing the patch to apply.
-		 * @return the revised list.
-		 * @throws PatchFailedException if the patch cannot be applied.
-		 */
-		public static <T> List<T> patch(List<? extends T> original, Patch<T> patch) throws PatchFailedException {
-				return patch.applyTo(original);
-		}
-
-		/**
-		 * Applies the given patch to the revised list and returns the original list.
-		 *
-		 * @param revised a {@link List} representing the revised list.
-		 * @param patch a {@link Patch} representing the patch to apply.
-		 * @return the original list.
-		 * @throws PatchFailedException if the patch cannot be applied.
-		 */
-		public static <T> List<T> unpatch(List<? extends T> revised, Patch<T> patch) {
-				return patch.restore(revised);
-		}
-
-		private static List<String> compressLines(List<String> lines, String delimiter) {
-				if (lines.isEmpty()) {
-						return Collections.emptyList();
-				}
-				return Collections.singletonList(String.join(delimiter, lines));
 		}
 
 		private DiffUtils() {}
